Add TMap, restructure internal lib
