+++ typerep-map.cabal
@@ -24,26 +24,26 @@ source-repository head
 library
   hs-source-dirs:      src
   exposed-modules:     Data.TypeRep.Map
+                       Data.TypeRep.TMap
   ghc-options:         -Wall
   build-depends:       base
-                     , typerep-map-internal
+                     , containers
+                     , ghc-prim
+                     , primitive
   default-extensions:  OverloadedStrings
                        RecordWildCards
                        ScopedTypeVariables
                        TypeApplications
   default-language:    Haskell2010
 
-library typerep-map-internal
-  hs-source-dirs:      internal
+library typerep-extra-impls
+  hs-source-dirs:      typerep-extra-impls
   exposed-modules:     Data.TypeRep.CMap
-                       Data.TypeRep.CacheMap
                        Data.TypeRep.OptimalVector
                        Data.TypeRep.Vector
   ghc-options:         -Wall
   build-depends:       base
                      , containers
-                     , ghc-prim
-                     , primitive
                      , vector
   default-extensions:  OverloadedStrings
                        RecordWildCards
@@ -65,7 +65,7 @@ test-suite typerep-map-test
                      , ghc-typelits-knownnat
                      , hedgehog
                      , typerep-map
-                     , typerep-map-internal
+                     , typerep-extra-impls
                      , tasty
                      , tasty-discover >= 4.1.1
                      , tasty-hedgehog
@@ -93,7 +93,8 @@ benchmark typerep-map-benchmark
                      , dependent-map
                      , dependent-sum
                      , ghc-typelits-knownnat
-                     , typerep-map-internal
+                     , typerep-map
+                     , typerep-extra-impls
   default-extensions:  OverloadedStrings
                        RecordWildCards
                        ScopedTypeVariables
